class rename general --> common
