+++ docs/commands/groups/Remove-TssGroupMember.md
@@ -108,7 +108,7 @@ This cmdlet supports the common parameters: -Debug, -ErrorAction, -ErrorVariable
 
 ## OUTPUTS
 
-### Thycotic.PowerShell.General.Delete
+### Thycotic.PowerShell.Common.Delete
 ## NOTES
 Requires TssSession object returned by New-TssSession
