+++ docs/commands/sites/Get-TssSite.md
@@ -61,7 +61,7 @@ This cmdlet supports the common parameters: -Debug, -ErrorAction, -ErrorVariable
 
 ## OUTPUTS
 
-### Thycotic.PowerShell.General.Site
+### Thycotic.PowerShell.Common.Site
 ## NOTES
 Requires TssSession object returned by New-TssSession
