+++ docs/commands/users/Remove-TssUserPii.md
@@ -92,7 +92,7 @@ This cmdlet supports the common parameters: -Debug, -ErrorAction, -ErrorVariable
 
 ## OUTPUTS
 
-### Thycotic.PowerShell.General.Delete
+### Thycotic.PowerShell.Common.Delete
 ## NOTES
 Requires TssSession object returned by New-TssSession
