+++ src/Thycotic.SecretServer/classes/common/Delete.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using System.Management.Automation;
 using System.Management.Automation.Runspaces;
 
-namespace Thycotic.PowerShell.General
+namespace Thycotic.PowerShell.Common
 {
     public class Delete
     {
