+++ src/Thycotic.SecretServer/classes/common/RequestStatus.cs
@@ -11,5 +11,6 @@ namespace Thycotic.PowerShell.Common
         public int StatusCode { get; set; }
         public string StatusDescription { get; set; }
         public string ResponseStatus { get; set; }
+        public string ResponseUri { get; set; }
     }
 }
