+++ src/functions/folders/Remove-TssFolderTemplate.ps1
@@ -22,7 +22,7 @@ function Remove-TssFolderTemplate {
     Requires TssSession object returned by New-TssSession
     #>
     [CmdletBinding(SupportsShouldProcess)]
-    [OutputType('Thycotic.PowerShell.General.Delete')]
+    [OutputType('Thycotic.PowerShell.Common.Delete')]
     param (
         # TssSession object created by New-TssSession for authentication
         [Parameter(Mandatory, ValueFromPipeline, Position = 0)]
@@ -67,7 +67,7 @@ function Remove-TssFolderTemplate {
                 }
 
                 if ($restResponse) {
-                    [Thycotic.PowerShell.General.Delete]$restResponse
+                    [Thycotic.PowerShell.Common.Delete]$restResponse
                 }
             }
         } else {
