+++ src/functions/groups/Remove-TssGroupMember.ps1
@@ -22,7 +22,7 @@ function Remove-TssGroupMember {
     Requires TssSession object returned by New-TssSession
     #>
     [CmdletBinding(SupportsShouldProcess)]
-    [OutputType('Thycotic.PowerShell.General.Delete')]
+    [OutputType('Thycotic.PowerShell.Common.Delete')]
     param (
         # TssSession object created by New-TssSession for authentication
         [Parameter(Mandatory, ValueFromPipeline, Position = 0)]
@@ -64,7 +64,7 @@ function Remove-TssGroupMember {
             }
 
             if ($restResponse) {
-                [Thycotic.PowerShell.General.Delete]@{
+                [Thycotic.PowerShell.Common.Delete]@{
                     Id         = $restResponse.id
                     ObjectType = $restResponse.objectType
                 }
