+++ src/functions/reports/Remove-TssReport.ps1
@@ -22,7 +22,7 @@ function Remove-TssReport {
     Requires TssSession object returned by New-TssSession
     #>
     [CmdletBinding(SupportsShouldProcess)]
-    [OutputType('Thycotic.PowerShell.General.Delete')]
+    [OutputType('Thycotic.PowerShell.Common.Delete')]
     param (
         # TssSession object created by New-TssSession for authentication
         [Parameter(Mandatory,ValueFromPipeline,Position = 0)]
@@ -60,7 +60,7 @@ function Remove-TssReport {
                 }
 
                 if ($restResponse) {
-                    [Thycotic.PowerShell.General.Delete]@{
+                    [Thycotic.PowerShell.Common.Delete]@{
                         Id = $report
                         ObjectType = 'Report'
                     }
