+++ src/functions/secret-dependencies/Remove-TssSecretDependency.ps1
@@ -22,7 +22,7 @@ function Remove-TssSecretDependency {
     Requires TssSession object returned by New-TssSession
     #>
     [CmdletBinding(SupportsShouldProcess, ConfirmImpact = 'High')]
-    [OutputType('Thycotic.PowerShell.General.Delete')]
+    [OutputType('Thycotic.PowerShell.Common.Delete')]
     param (
         # TssSession object created by New-TssSession for authentication
         [Parameter(Mandatory, ValueFromPipeline, Position = 0)]
@@ -61,7 +61,7 @@ function Remove-TssSecretDependency {
                 }
 
                 if ($restResponse) {
-                    [Thycotic.PowerShell.General.Delete]@{
+                    [Thycotic.PowerShell.Common.Delete]@{
                         Id         = $restResponse.id
                         ObjectType = $restResponse.objectType
                     }
