+++ src/functions/secret-permissions/Remove-TssSecretPermission.ps1
@@ -22,7 +22,7 @@ function Remove-TssSecretPermission {
     Requires TssSession object returned by New-TssSession
     #>
     [CmdletBinding(SupportsShouldProcess)]
-    [OutputType('Thycotic.PowerShell.General.Delete')]
+    [OutputType('Thycotic.PowerShell.Common.Delete')]
     param (
         # TssSession object created by New-TssSession for authentication
         [Parameter(Mandatory,ValueFromPipeline,Position = 0)]
@@ -60,7 +60,7 @@ function Remove-TssSecretPermission {
                 }
 
                 if ($restResponse) {
-                    [Thycotic.PowerShell.General.Delete]@{
+                    [Thycotic.PowerShell.Common.Delete]@{
                         Id = $restResponse.id
                         ObjectType = $restResponse.objectType
                     }
