+++ src/functions/secrets/Remove-TssSecret.ps1
@@ -22,7 +22,7 @@ function Remove-TssSecret {
     Requires TssSession object returned by New-TssSession
     #>
     [cmdletbinding(SupportsShouldProcess)]
-    [OutputType('Thycotic.PowerShell.General.Delete')]
+    [OutputType('Thycotic.PowerShell.Common.Delete')]
     param(
         # TssSession object created by New-TssSession for authentication
         [Parameter(Mandatory, ValueFromPipeline, Position = 0)]
@@ -63,7 +63,7 @@ function Remove-TssSecret {
                 }
 
                 if ($restResponse) {
-                    [Thycotic.PowerShell.General.Delete]$restResponse
+                    [Thycotic.PowerShell.Common.Delete]$restResponse
                 }
             }
         } else {
